fix: eliminate Double round-trip in picker conversion preview + fix tests
- MainViewModel.getCurrentBaseValueAsNumber and getResultAsNumber now
  return LiveData<BigDecimal> instead of LiveData<Double>. Previously the
  BigDecimal string value was converted to Double just to be converted
  back to BigDecimal in buildConversionText, losing precision.
- SearchableSpinner{,Dialog,DialogAdapter}.setCurrentSum now accepts
  BigDecimal instead of Double.
- SearchableSpinnerDialogAdapter.buildConversionText uses the incoming
  BigDecimal directly (no BigDecimal(double) constructor).
- MathUtilsTest updated to use BigDecimal signatures. Dropped the two
  Float.POSITIVE_INFINITY / NEGATIVE_INFINITY cases since BigDecimal
  has no infinity representation; added a BigDecimal.ZERO divisor case.

Remaining Float usages are all intentional library boundaries:
ChartAdapter (SparkAdapter API), Database legacy fee migration read,
DimenUtils (dp/px), and generic SharedPreferenceFloatLiveData utility.

+++ app/src/main/kotlin/de/salomax/currencies/view/main/spinner/SearchableSpinner.kt
@@ -11,6 +11,7 @@ import androidx.fragment.app.FragmentActivity
 import de.salomax.currencies.R
 import de.salomax.currencies.model.Currency
 import de.salomax.currencies.model.Rate
+import java.math.BigDecimal
 
 class SearchableSpinner : AppCompatSpinner {
 
@@ -88,7 +89,7 @@ class SearchableSpinner : AppCompatSpinner {
         // set in dialog
         spinnerDialog.setCurrentRate(currentRate)
     }
-    fun setCurrentSum(currentSum: Double) {
+    fun setCurrentSum(currentSum: BigDecimal) {
         // set in dialog
         spinnerDialog.setCurrentSum(currentSum)
     }
